requirements.txt
numpy
opencv-python
tensorflow
streamlit
Pillow
pandas